Converting to maven project
