+++ src/main/java/com/jfsplit/core/FileJoinProcess.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package jfsplit.core;
+package com.jfsplit.core;
 
 import java.io.File;
 import java.io.IOException;
