+++ src/main/java/com/jfsplit/core/FileOperations.java
@@ -19,15 +19,13 @@
  * 
  */
 
-package jfsplit.core;
+package com.jfsplit.core;
 
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.io.IOException;
-
 import java.nio.channels.FileChannel;
-
 import java.text.DecimalFormat;
 
 /**
