+++ src/main/java/com/jfsplit/core/FileProcess.java
@@ -18,7 +18,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package jfsplit.core;
+package com.jfsplit.core;
 
 /**
  * The <code>FileProcess</code> interface specifies the methods a
