+++ src/main/java/com/jfsplit/gui/FileChecksumPanel.java
@@ -19,28 +19,16 @@
  * 
  */
 
-package jfsplit.gui;
+package com.jfsplit.gui;
 
+import com.jfsplit.core.*;
+import net.miginfocom.swing.MigLayout;
+
+import javax.swing.*;
+import javax.swing.border.TitledBorder;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.io.File;
-
-import javax.swing.JButton;
-import javax.swing.JFileChooser;
-import javax.swing.JLabel;
-import javax.swing.JOptionPane;
-import javax.swing.JPanel;
-import javax.swing.JProgressBar;
-import javax.swing.JTextField;
-import javax.swing.border.TitledBorder;
-
-import net.miginfocom.swing.MigLayout;
-
-import jfsplit.core.FileChecksumProcess;
-import jfsplit.core.FileOperations;
-import jfsplit.core.FileProcess;
-import jfsplit.core.ProcessCaller;
-import jfsplit.core.Status;
 
 /**
  * Panel which contains required gui elements for checksum calculator
